refactor: 식재료 가격 타입 Integer로 변경

+++ src/main/java/org/springeel/oneclickrecipe/domain/recipefood/dto/service/RecipeFoodReadResponseDto.java
@@ -5,7 +5,7 @@ import org.springeel.oneclickrecipe.domain.food.entity.UnitType;
 public record RecipeFoodReadResponseDto(
     String name,
     String amount,
-    Float price,
+    Integer price,
     UnitType unit
 ) {
